refactor(utils): 更新 load_check_run 模块导入
- 从 tool 模块中添加 toMainUi 函数的导入
- 保持原有其他导入功能不变

refactor(AutoPlan): 重构检查运行工具类结构

- 添加Base抽象基类并定义build和run静态方法
- Domain类继承Base基类
- LeyLineOutcrop类继承Base基类
- StygianOnslaught类继承Base基类
- 统一了类的继承结构和方法定义

refactor(AutoPlan): 重构自动任务模块代码结构

- 将秘境、地脉花、幽境危战任务处理逻辑封装为独立类
- 提取构建方法为各任务类的静态build方法
- 添加运行方法为各任务类的静态run方法
- 新增autoRunList函数统一处理多种任务类型的执行
- 优化代码组织结构提升可维护性

refactor(AutoPlan): 重构运行配置解析逻辑

- 提取订单构建逻辑到独立的 buildOrder 函数中
- 简化 load_check_run.js 中的配置处理流程
- 添加 uid 和 bgi_tools 模块的导入依赖
- 优化代码结构提高可读性和维护性

refactor(AutoPlan): 重构秘境配置加载逻辑

- 提取秘境、地脉和幽境构建逻辑到独立函数中
- 优化代码结构提高可读性和维护性
- 添加多天执行计划支持功能
- 改进配置加载方式支持多种数据源
- 优化执行顺序初始化流程
- 统一错误处理和日志记录机制

refactor(AutoPlan): 重构幽境危战检查功能模块

- 将 checkAndFilterStygianOnslaught 函数从 config.js 移动到新的 load_check_run.js 文件
- 重命名文件 load_and_run.js 为 load_check_run.js 以更好地反映其功能
- 更新 main.js 中的导入路径，指向新的工具函数文件
- 在 config.js 中移除幽境危战相关的检查函数代码
- 保持原有功能不变，仅进行代码组织结构优化

feat(AutoPlan): 添加 LoadType 类型并导入 getDayOfWeek 工具函数

- 从 config 模块导入 LoadType 类型定义
- 从 tool 模块导入 getDayOfWeek 工具函数
- 更新 import 语句以包含新增的功能模块

refactor(config): 将加载配置功能从config.js迁移到新的load_and_run.js文件

- 将loadMode和initRunOrderList函数从config.js移动到load_and_run.js
- 重命名run.js为load_and_run.js以更好地反映其功能
- 更新main.js中的导入路径，指向新的load_and_run.js文件
- 移除config.js中已迁移的加载相关函数代码
- 保持原有功能不变，仅进行代码组织结构优化

refactor(config): 重构幽境危战任务检查功能

- 将 checkAndFilterStygianOnslaught 函数从 main.js 移至 config.js 文件
- 在 config.js 中新增 activity 工具导入和相关检查函数
- 添加代码分隔注释以提高配置文件结构清晰度
- 更新 main.js 导入语句以包含新的检查函数
- 移除 main.js 中重复的函数定义实现
- 优化代码组织结构提升模块化程度

refactor(config): 重构配置模块并导出核心功能函数

- 将 config、LoadType、LoadMap 等对象改为可导出
- 新增多复选框映射表相关函数 getMultiCheckboxMap 和 getValueByMultiCheckboxName
- 添加密钥验证函数 checkKey
- 导出 initSettings 函数供外部使用
- 新增 loadMode 函数处理不同加载方式的配置
- 实现 initRunOrderList 函数初始化执行顺序列表
- 在 main.js 中导入并使用新增的导出函数
- 移除 main.js 中重复的函数实现代码
